+++ docs/conf.py
@@ -30,7 +30,7 @@ version = ''
 # The full version, including alpha/beta/rc tags.
 release = ''
 try:
-    from qiskit_iqm import __version__ as version
+    from iqm.qiskit_iqm import __version__ as version
 except ImportError:
     pass
 else:
