+++ src/iqm/qiskit_iqm/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 """Qiskit adapter for IQM's quantum computers.
 """
-from qiskit_iqm.fake_backends import IQMErrorProfile, IQMFakeAdonis
-from qiskit_iqm.fake_backends.iqm_fake_backend import IQMFakeBackend
-from qiskit_iqm.iqm_job import IQMJob
-from qiskit_iqm.iqm_provider import IQMBackend, IQMProvider, __version__
+from iqm.qiskit_iqm.fake_backends import IQMErrorProfile, IQMFakeAdonis
+from iqm.qiskit_iqm.fake_backends.iqm_fake_backend import IQMFakeBackend
+from iqm.qiskit_iqm.iqm_job import IQMJob
+from iqm.qiskit_iqm.iqm_provider import IQMBackend, IQMProvider, __version__
