+++ src/iqm/qiskit_iqm/iqm_backend.py
@@ -19,7 +19,7 @@ from abc import ABC
 import re
 from typing import Final, Optional
 
-from iqm_client import QuantumArchitectureSpecification
+from iqm.iqm_client import QuantumArchitectureSpecification
 from qiskit.circuit import Parameter
 from qiskit.circuit.library import CZGate, IGate, Measure, RGate
 from qiskit.providers import BackendV2
