+++ src/iqm/qiskit_iqm/iqm_provider.py
@@ -19,16 +19,16 @@ from typing import Optional, Union
 from uuid import UUID
 import warnings
 
-from iqm_client import Circuit, HeraldingMode, Instruction, IQMClient
-from iqm_client.util import to_json_dict
+from iqm.iqm_client import Circuit, HeraldingMode, Instruction, IQMClient
+from iqm.iqm_client.util import to_json_dict
 import numpy as np
 from qiskit import QuantumCircuit
 from qiskit.providers import JobStatus, JobV1, Options
 
-from qiskit_iqm.fake_backends import IQMFakeAdonis
-from qiskit_iqm.iqm_backend import IQMBackendBase
-from qiskit_iqm.iqm_job import IQMJob
-from qiskit_iqm.qiskit_to_iqm import MeasurementKey
+from iqm.qiskit_iqm.fake_backends import IQMFakeAdonis
+from iqm.qiskit_iqm.iqm_backend import IQMBackendBase
+from iqm.qiskit_iqm.iqm_job import IQMJob
+from iqm.qiskit_iqm.qiskit_to_iqm import MeasurementKey
 
 try:
     __version__ = version('qiskit-iqm')
