+++ tests/fake_backends/test_fake_adonis.py
@@ -16,7 +16,7 @@
 """
 from qiskit_aer.noise.noise_model import NoiseModel
 
-from qiskit_iqm import IQMFakeAdonis
+from iqm.qiskit_iqm import IQMFakeAdonis
 
 
 def test_iqm_fake_adonis():
