+++ tests/test_iqm_provider.py
@@ -17,7 +17,7 @@
 from importlib.metadata import version
 import uuid
 
-from iqm_client import HeraldingMode, IQMClient, RunResult, RunStatus
+from iqm.iqm_client import HeraldingMode, IQMClient, RunResult, RunStatus
 from mockito import ANY, mock, patch, when
 import numpy as np
 import pytest
@@ -26,8 +26,8 @@ from qiskit.circuit import Parameter
 from qiskit.circuit.library import RGate
 from qiskit.compiler import transpile
 
-from qiskit_iqm import IQMBackend, IQMJob, IQMProvider
-from qiskit_iqm.iqm_provider import IQMFacadeBackend
+from iqm.qiskit_iqm import IQMBackend, IQMJob, IQMProvider
+from iqm.qiskit_iqm.iqm_provider import IQMFacadeBackend
 
 
 @pytest.fixture
