+++ tox.ini
@@ -23,11 +23,9 @@ extras =
     testing
 commands =
     python -m black --check src tests
-    python -m pytest tests --verbose --cov --cov-report term-missing --junitxml=test_report.xml --doctest-modules src
-    python -m pytest --pylint --mypy src/ --verbose --strict
-    python -m pytest --pylint --mypy tests/ --pylint-rcfile=tests/.pylintrc --verbose --strict
-    pytest --mypy tests/ --verbose --strict
-    pytest --isort tests/ src/ --verbose
+    python -m mypy -p iqm.qiskit_iqm
+    python -m pytest --doctest-modules --isort --pylint src --verbose --strict
+    python -m pytest --cov iqm.qiskit_iqm --cov-report=term-missing --junitxml=test_report.xml --isort --doctest-modules --mypy --pylint --pylint-rcfile=tests/.pylintrc tests --verbose --strict
 
 [testenv:format]
 description =
